Moving files now works

+++ Move.py
@@ -4,11 +4,17 @@ import sys
 import shutil
 import os
 
-def main(playlist, mus_fol, and_fol, name):
+def main(playlist, mus_fol, and_fol):
+
+    if mus_fol.endswith("/"):
+        mus_fol[:-1]
+    if and_fol.endswith("/"):
+        and_fol[:-1]
 
     mus_fol = os.path.expanduser(mus_fol)
     and_fol = os.path.expanduser(and_fol)
     playlist = os.path.expanduser(playlist)
+    name = str(playlist)
     songs = []
     song_fs = []
 
@@ -16,11 +22,18 @@ def main(playlist, mus_fol, and_fol, name):
         songs = fin.readlines()
 
     play_dir = and_fol + '/' + name
+    
     if not os.path.exists(play_dir):
         os.mkdir(play_dir)
     os.chdir(play_dir)
 
     for song in songs:
+        
+        # Remove \n from songs list
+        if song.endswith("\n"):
+            song[:-2]
+
+
         """
         song.replace(" ","\ ")
         song.replace("(","\(")
@@ -28,27 +41,43 @@ def main(playlist, mus_fol, and_fol, name):
         song.replace("[","\]")
         song.replace("]","\]")
         """
-        fold,song_f = song.split('/',1)
-        song_fs.append(song_f)
-        if not os.path.isdir(fold):
-            os.mkdir(fold)
+        
+        artist,other = song.split('/',1)
+        album,song_fs= other.split('/',1)
+        
+        
+        and_loc_artist = and_fol + artist
+        and_loc_album = and_loc_artist + '/' + album
+        and_loc_song = and_loc_album + '/' + song_fs
+ 
 
-        if not os.path.exists(song):
-            shutil.copyfile(mus_fol + '/' + song,'./' + song)    
+        if not os.path.isdir(and_loc_artist):
+            os.mkdir(and_loc_artist)
 
-        print(song_f + " has been copied")
+        if not os.path.isdir(and_loc_album):
+            os.mkdir(and_loc_album)
 
+        if not os.path.exists(and_loc_song):
+            print (mus_fol)
+            songloc = mus_fol + "/" + artist + "/" + album + "/" + song_fs 
+            shutil.copyfile(songloc[:-1], and_loc_song)    
+            print(songloc + " has been copied")
+    
     print("deleting stuff")
+
+    music_formats = [".mp3",".flac"]
+
     for dirpath, dirnames, song_f in os.walk('./'): 
         if len(song_f) == 1:
             if not song_f in song_fs:
-                # notifiaction message here
-                os.remove(dirpath + '/' + song_f)
+                for formats in music_formats:
+                    if str(song_f).endswith(formats):
+                        print(song_f)
+                        os.remove(dirpath + '/' + song_f)
 
 
 if __name__=="__main__":
     print("The playlist file is " + sys.argv[1])
     print("The music root folder is " + sys.argv[2])
     print("The android root folder is " + sys.argv[3])
-    print("The playlist name is " + sys.argv[4])
-    main(sys.argv[1],sys.argv[2],sys.argv[3],sys.argv[4])
+    main(sys.argv[1],sys.argv[2],sys.argv[3])
